classify_webbies/requirements.txt
eventlet
dnspython
netaddr
